Change: In media info tooltip high bit rate displayed in Mb/s  Change: Get packets timeout increased from 60 seconds to 1800 seconds.  Bugfix: Program crashes on opening file with very high bitrate (I hope I fixed it this time)  Bugfix: Timeout do not work while waiting for external program (ffprobe)

+++ FFProbe.cs
@@ -185,7 +185,7 @@ namespace FFBitrateViewer
     public class FFProbeStream
     {
         [JsonProperty("bit_rate")]
-        public int? BitRate { get; set; }
+        public long? BitRate { get; set; }
 
         [JsonProperty("bits_per_sample")]
         public int? BitsPerSample { get; set; }
